Updates and Hotfixes

+++ src/main/java/pl/petersen/cinemamanager/service/HallService.java
@@ -1,8 +1,11 @@
 package pl.petersen.cinemamanager.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import pl.petersen.cinemamanager.entity.Hall;
+import pl.petersen.cinemamanager.entity.Reservation;
 import pl.petersen.cinemamanager.entity.Seat;
 import pl.petersen.cinemamanager.repository.HallRepository;
 import pl.petersen.cinemamanager.repository.SeatRepository;
@@ -17,13 +20,33 @@ public class HallService {
 
     private final HallRepository hallRepository;
     private final SeatRepository seatRepository;
+    private final SeanceService seanceService;
+    private final ReservationService reservationService;
 
     @Autowired
-    public HallService(HallRepository hallRepository, SeatRepository seatRepository) {
+    public HallService(HallRepository hallRepository, SeatRepository seatRepository, SeanceService seanceService, ReservationService reservationService) {
         this.hallRepository = hallRepository;
         this.seatRepository = seatRepository;
+        this.seanceService = seanceService;
+        this.reservationService = reservationService;
     }
 
+
+    public Boolean doesHallContainAnyReservations(Hall hall) {
+        if (hall.getId() == null) {
+            return false;
+        }
+        Optional<Hall> oldHall = hallRepository.findById(hall.getId());
+        List<Reservation> allReservations = reservationService.findAll();
+        if (oldHall.isPresent()) {
+            if (allReservations.stream()
+                    .anyMatch(r -> r.getSeance().getHall().equals(oldHall.get()))) {
+                return hall.getNumOfSeatsPerRow() < oldHall.get().getNumOfSeatsPerRow()
+                        || hall.getRows() < oldHall.get().getRows();
+            }
+        }
+        return false;
+    }
 
     public Optional<Hall> findById(Long hallId) {
         return hallRepository.findById(hallId);
@@ -56,8 +79,8 @@ public class HallService {
 
 
     private Seat checkSeatExistence(Character letter, Integer num) {
-        if(seatRepository.count() > 0) {
-        List<Seat> existingSeats = getSeats();
+        if (seatRepository.count() > 0) {
+            List<Seat> existingSeats = getSeats();
             for (Seat existingSeat : existingSeats) {
                 if (existingSeat.getRow().equals(letter)
                         && existingSeat.getNumber().equals(num)) {
@@ -74,7 +97,12 @@ public class HallService {
         return hallRepository.findAll();
     }
 
-    public void deleteById(Long deleteId) {
-        hallRepository.deleteById(deleteId);
+    public Boolean deleteById(Long deleteId) {
+        long count = seanceService.countByHallId(deleteId);
+        if (count == 0) {
+            hallRepository.deleteById(deleteId);
+            return true;
+        }
+        return false;
     }
 }
